requirements.txt
altair==4.2.0
attrs==22.1.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
blinker==1.5
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
crashtest==0.3.1
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
dulwich==0.20.50
entrypoints==0.4
filelock==3.8.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core
google-auth==2.15.0
google-cloud-core
google-cloud-firestore
grpcio
grpcio-status
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.11.0
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
msgpack==1.0.4
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.9.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
plotly==5.11.0
proto-plus==1.22.1
ptyprocess==0.7.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
pylev==1.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.6.0
rsa==4.9
semver==2.13.0
setuptools-scm==7.0.5
shellingham==1.5.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.2
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.0
watchdog==2.2.0
webencodings==0.5.1
xattr==0.9.9
zipp==3.11.0
protobuf